Last step of UI creation complete
So, the UI creation is over. I saved the files and am updating them here but could not run them due to some issues with my pc. I will run them soon and then update if there are any flaws, otherwise this is the final file updation from my side.